Relocate some code
