+++ internal/cluster/distribution/eks/defaults.go
